backend-gateway/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
python-multipart
requests
pypdf
python-docx
reportlab
